RecognitionMonitor is now an interface, it looks like everything works fine
